requirements.txt
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
blis @ file:///C:/ci/cython-blis_1594248830923/work
bpemb==0.3.2
brotlipy==0.7.0
catalogue==1.0.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1606255207413/work
chardet @ file:///C:/ci/chardet_1607706910910/work
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
convertdate==2.3.0
coverage==5.3.1
cryptography @ file:///C:/ci/cryptography_1607639129468/work
cycler==0.10.0
cymem @ file:///C:/ci/cymem_1605062363169/work
Cython==0.29.14
decorator==4.4.2
Deprecated==1.2.10
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.0.12
flair==0.7
Flask==1.1.2
ftfy==5.8
future==0.18.2
gdown==3.12.2
gensim==3.8.3
geocoder==1.38.1
graphviz==0.15
gunicorn==20.0.4
holidays==0.10.2
hyperopt==0.2.5
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Janome==0.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.0.2
kiwisolver==1.3.1
konoha==4.6.2
korean-lunar-calendar==0.2.1
langdetect==1.0.8
lxml==4.6.2
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib==3.2.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mpld3==0.3
murmurhash @ file:///C:/ci/murmurhash_1607438161064/work
networkx==2.5
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.0
overrides==3.0.0
packaging==20.8
pandas==1.1.0
parso==0.7.1
patsy==0.5.1
pickleshare==0.7.5
plac @ file:///C:/ci/plac_1594260069687/work
preshed==3.0.2
prompt-toolkit==3.0.8
protobuf==3.14.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.7.3
PyMeeus==0.3.7
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
PySocks @ file:///C:/ci/pysocks_1594394709107/work
pytextrank==2.0.3
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
PyYAML==5.3.1
ratelim==0.1.6
regex==2020.11.13
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.0
segtok==1.5.10
sentencepiece==0.1.91
six @ file:///C:/ci/six_1605205426665/work
smart-open==4.0.1
spacy @ file:///C:/ci/spacy_1608321173027/work
sqlitedict==1.7.0
srsly @ file:///C:/ci/srsly_1607530581965/work
statsmodels==0.11.1
tabulate==0.8.7
thinc @ file:///C:/ci/thinc_1607710279177/work
threadpoolctl==2.1.0
tokenizers==0.9.3
torch==1.7.1
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets==5.0.5
transformers==3.5.1
tweepy==3.9.0
typing-extensions==3.7.4.3
u8darts==0.5.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1606938623459/work
wasabi @ file:///tmp/build/80754af9/wasabi_1599244993840/work
wcwidth==0.2.5
Werkzeug==1.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
